changed addschoolyear method in controller to editschoolyear method
If the referrer isn't the addscoolyear page then the user is redirected to their referrer if there were errors
failed requests from addschoolyear get the page returned with errors on page
successful requests are always redirected to referrer

Added a bunch to the school page. Getting the school's schedule put on the page. Still need to add class time blocks, the form for grading periods, and datepickers for all date inputs

Signed-off-by: kevinw <kevinwhite88@gmail.com>