I thought DozerBeanMappre was an interface, changed to the Mapper interface in all service layer classes
Added District object, we'll need this for reporting and it could open the doors for multi-tenancy

Still working on Events for the calendar, I now have it querying the school event tables

I added a test, but it's not working correctly. I think I need to implement hashcodes and equals in all my Beans to get it to work. I believe the equals() method is being called but since I instantiate a new object in the method call I cna't get the generic equals to work for assertions.

Signed-off-by: kevinw <kevinwhite88@gmail.com>