more #16 and #17
I think I like where these are for now. There's not a way to delete anything yet, but everything else is fine.

Subjects can be created from anywhere, but you gotta pick which school to create a subject for (will expand this later so you can select multiple/all schools)

Can only delete on the back end but I'll add that stuff later and wlil create new issues for that

Courses can only be created from the school page or on the subject's page

I think I have most of the basics covered for subjects and courses

Signed-off-by: kevinw <kevinwhite88@gmail.com>